Now, I think I live with these ansi-term keys
We will see how long terminals in emacs lasts this time.  Not sure what
to do with the hangs on OSX, but this helps

http://www.emacswiki.org/emacs/MacOSTweaks#toc19

Emacs on OSX is IMHO better than the other editors I’ve used on Mac, but it’s still annoyingly unstable. In particular, it has a nasty tendency to hang on bad bytecode and stalled network connections, and these hangs can’t be broken with a C-g. The solution I’ve found is to attach to it with GDB, then enter a recursive edit:

    $ gdb /Applications/Emacs.app/Contents/MacOS/Emacs `ps uxwww | grep [E]macs | awk '{ print $2 }'`
    ...
    (gdb) p Frecursive_edit()
then hit C-c. Finally

    (gdb) detach
    (gdb) quit